requirements.txt
numpy==1.24.1
pandas==1.5.2
flask==2.2.2
requests==2.28.1
python-telegram-bot==13.15
pyChatGPT==0.4.2
openai
google-api-python-client
google-cloud-texttospeech==2.13.0